// client/src/generatedapi.ts
export class CreateMessageRequest implements ICreateMessageRequest {
    content?: string | undefined;

    constructor(data?: ICreateMessageRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.content = _data["content"];
        }
    }

    static fromJS(data: any): CreateMessageRequest {
        data = typeof data === 'object' ? data : {};
        let result = new CreateMessageRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["content"] = this.content;
        return data;
    }
}

export interface ICreateMessageRequest {
    content?: string | undefined;
}
// ...
export class MessageDto implements IMessageDto {
    id?: string;
    room?: string | undefined;
    senderUsername?: string | undefined;
    type?: string | undefined;
    content?: string | undefined;
    recipientUserId?: string | undefined;
    sentAt?: Date;

    constructor(data?: IMessageDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.room = _data["room"];
            this.senderUsername = _data["senderUsername"];
            this.type = _data["type"];
            this.content = _data["content"];
            this.recipientUserId = _data["recipientUserId"];
            this.sentAt = _data["sentAt"] ? new Date(_data["sentAt"].toString()) : undefined as any;
        }
    }

    static fromJS(data: any): MessageDto {
        data = typeof data === 'object' ? data : {};
        let result = new MessageDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["room"] = this.room;
        data["senderUsername"] = this.senderUsername;
        data["type"] = this.type;
        data["content"] = this.content;
        data["recipientUserId"] = this.recipientUserId;
        data["sentAt"] = this.sentAt ? this.sentAt.toISOString() : undefined as any;
        return data;
    }
}

export interface IMessageDto {
    id?: string;
    room?: string | undefined;
    senderUsername?: string | undefined;
    type?: string | undefined;
    content?: string | undefined;
    recipientUserId?: string | undefined;
    sentAt?: Date;
}